Restructure to follow MVVM standards
